introduce new module and move simple classes over
