+++ legend-engine-core/legend-engine-core-identity/legend-engine-identity-core/src/main/java/org/finos/legend/engine/shared/core/identity/credential/AnonymousCredential.java
