+++ legend-engine-core/legend-engine-core-shared/legend-engine-shared-core/pom.xml
@@ -58,6 +58,10 @@
             <groupId>org.finos.legend.engine</groupId>
             <artifactId>legend-engine-protocol-pure</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.finos.legend.engine</groupId>
+            <artifactId>legend-engine-identity-core</artifactId>
+        </dependency>
         <!-- ENGINE -->
 
         <!-- JACKSON -->
